+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - allow RETURNING with empty result set (such as UPDATE).
 
+### Changed
+- fixed Number.MarshalJSON (see #112)'
+
 ### [2.8.2] - 2018-10-01
 ### Changed
 - Fix the driver.Valuer handling, make it the last resort